Fix inverse_binomial_coefficient().
This fixes a really ugly bug. inverse_binomial_coefficent was actually
returning the binomial coefficient, instead of its inverse. Luckily,
this function was only being used in exactly one place, in the
computation of IC1c, and IC1c only rarely needs to be computed.
Moveover, even though the answer from this function was quite wrong, it
didn't matter _that_ much because binomial_coefficient(j, 0) == 1 will
still be the same when inverted and, when it was being used for larger
input inverse_binomial_coefficient() was being multiplied by things that
were very very small. For these reasons, at least in most cases, the
error from this bug was quite small. Especially in our computations of
the zeta function I believe that the error will be quite small.

It is possible that there are some cases where the error is much
bigger, though.